+++ sc_bp_watcher.py
@@ -58,7 +58,7 @@ try:
 except ImportError:
     winsound = None
 
-__version__ = '3.5.0'
+__version__ = '3.5.1'
 
 
 def _mitgeliefert(name):